McuMonitor: Flash table to 16 bit.

+++ src/simulator/elements/processors/picprocessor.h
@@ -47,8 +47,8 @@ class PicProcessor : public BaseProcessor
 
         virtual uint8_t getRamValue( int address ) override;
         virtual void    setRamValue( int address, uint8_t value ) override;
-        virtual int  getFlashValue( int address ) override;
-        virtual void setFlashValue( int address, uint8_t value ) override;
+        virtual uint16_t getFlashValue( int address ) override;
+        virtual void setFlashValue( int address, uint16_t value ) override;
         virtual uint8_t getRomValue( int address ) override;
         virtual void    setRomValue( int address, uint8_t value ) override;
